Enforce Prime mining template ProofHash validity

+++ src/LLP/types/miner.h
@@ -653,7 +653,7 @@ namespace LLP
          *  CreateProducer() is skipped (~99% reduction in steady-state key ops).
          *
          *  Protected by MUTEX. **/
-        uint32_t   m_nCachedExtraNonce{0};
+        uint64_t   m_nCachedExtraNonce{0};
         uint1024_t m_hashLastExtraNonceTip;
         TAO::API::CredentialCache m_miningCredentialCache;
